:construction: Changed the new tests to be async

+++ jovo-cli/test/Helpers.ts
@@ -14,7 +14,7 @@ import * as path from 'path';
  * @param {string} waitText The text to wait for in the output
  * @returns
  */
-export async function runJovoCommand(command: string, parameters: string[], cwd: string, waitText: string) {
+export async function runJovoCommand(command: string, parameters: string[], cwd: string, waitText: string | null, errorText?: string) {
 	parameters.unshift(command);
 
 	// Get the script from the correct location depending on if we are in main or subfolder
@@ -29,10 +29,15 @@ export async function runJovoCommand(command: string, parameters: string[], cwd:
 			cwd,
 		});
 		child.stderr.on('data', (data) => {
+			if (errorText && data.toString().indexOf(errorText) > -1) {
+				child.kill();
+				return resolve(data.toString());
+			}
+
 			reject(new Error(data.toString()));
 		});
 		child.stdout.on('data', (data) => {
-			if (data.toString().indexOf(waitText) > -1) {
+			if (waitText !== null && data.toString().indexOf(waitText) > -1) {
 				child.kill();
 				resolve(data.toString());
 			}
